+++ bend.cabal
@@ -29,8 +29,10 @@ library
                       Core.Collapse,
                       Core.Deps,
                       Core.Equal,
-                      Core.EtaForm,
-                      Core.Flatten,
+                      Core.Adjust.EtaForm,
+                      Core.Adjust.FlattenPats,
+                      Core.Adjust.DesugarPats,
+                      Core.Adjust.DesugarFrks,
                       Core.FreeVars,
                       Core.Import,
                       Core.LHS,
